Moved file handling logic to SplitStoreService
