+++ core/src/main/java/bisq/core/account/witness/AccountAgeWitnessStore.java
@@ -18,17 +18,13 @@
 package bisq.core.account.witness;
 
 import bisq.network.p2p.storage.P2PDataStorage;
-import bisq.network.p2p.storage.payload.PersistableNetworkPayload;
 import bisq.network.p2p.storage.persistence.SplitStore;
 
 import com.google.protobuf.Message;
 
 import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
-import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
 
@@ -39,9 +35,6 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 public class AccountAgeWitnessStore extends SplitStore {
-    @Getter
-    private Map<P2PDataStorage.ByteArray, PersistableNetworkPayload> map = new ConcurrentHashMap<>();
-
     AccountAgeWitnessStore() {
     }
